+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - omitting error message for failing MCP server that are not enabled
 - sub-agent tool presentation and context persistence improvements
 - added icon for auto-approve to agent selector for immediate recognition
+- showing streaming of reasoning (thinking) on agent messages
 
 ## [0.19.0]